Different way to use alpha instead of blend function
Blend functions are too overkill, and using it cause a lot of
complication when I add multiple objects that need obverlapped.

So, this approach from
[LearOpenGL](https://learnopengl.com/Advanced-OpenGL/Blending) works
perfectly and output is what I want.

To see how it's work this commit should be merge with the commit where I
add multiple tree object into the scene.